% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeDataBuild.R
\name{addAtlasCohort}
\alias{addAtlasCohort}
\title{Add custom atlas cohorts to the results dataset. Just needs an atlasId and cdm and the cohort will be generated with the
result being added to the final dataset.}
\usage{
addAtlasCohort(
  atlasId,
  configFilePath = "config/global-cfg.yml",
  removeExisting = FALSE,
  dataSources = NULL,
  logFileName = "rbAtlasCohort.log",
  dataDir = "data",
  addResult = TRUE
)
}
\arguments{
\item{atlasId}{add specified atlas id}

\item{configFilePath}{path on disk to a yaml configuration file containing all needed config setting to point at relevant cdm}

\item{removeExisting}{removes any already existing data (useful if you change the cohort definition in atlas as rewardb does not track definition verions)}

\item{dataSources}{vector of strings or null - keys to cdm stores to use. By default all cdms are added for a given config file}

\item{logFileName}{location to store parallel logger logfile}
}
\description{
Add custom atlas cohorts to the results dataset. Just needs an atlasId and cdm and the cohort will be generated with the
result being added to the final dataset.
}
