% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildCohorts.R
\name{addAtlasOutcomeCohort}
\alias{addAtlasOutcomeCohort}
\title{Pulls cohort from atlas and then creates it in the rewardb outcome cohorts table}
\usage{
addAtlasOutcomeCohort(connection, config, atlasId, dataSources)
}
\arguments{
\item{connection}{DatabaseConnector connection to cdm}

\item{atlasId}{atlasId in WebAPI}

\item{dataSources}{dataSources to run cohort on}
}
\description{
Pulls cohort from atlas and then creates it in the rewardb outcome cohorts table
}
