% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildDashboard.R
\name{addCemEvidence}
\alias{addCemEvidence}
\title{Add CEM based associations}
\usage{
addCemEvidence(appContext, connection)
}
\arguments{
\item{appContext}{application context loaded from yaml}

\item{connection}{DatabaseConnector connection to cdm}
}
\description{
This is used for the automated construction of negative control sets and the indication labels
}
