% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{addManualNegativeOutcomeControls}
\alias{addManualNegativeOutcomeControls}
\title{Input should specify a csv file with the columns target_concept_id and outcome_concept_id
This function will then work out which cohorts they should apply to as it is done on the concept level
This is not needed if automated manual controls are specified from the cem}
\usage{
addManualNegativeOutcomeControls(appContext)
}
\arguments{
\item{appContext}{takes a rewardb application context}
}
\description{
Input should specify a csv file with the columns target_concept_id and outcome_concept_id
This function will then work out which cohorts they should apply to as it is done on the concept level
This is not needed if automated manual controls are specified from the cem
}
