% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOutcomeSummary.R
\name{addOutcomeSummary}
\alias{addOutcomeSummary}
\title{Add outcome summary resultS}
\usage{
addOutcomeSummary(connection, config, outcomeCohortIds, dataSources)
}
\arguments{
\item{connection}{DatabaseConnector connection object}
}
\description{
Add outcome summary resultS
}
