% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Atlas.R
\name{addPhenotypeLibrary}
\alias{addPhenotypeLibrary}
\title{add Phenotype Library Cohorts}
\usage{
addPhenotypeLibrary(
  connection,
  config,
  libraryRepo = "OHDSI/PhenotypeLibrary",
  ref = "develop",
  local = FALSE,
  packageName = "PhenotypeLibrary",
  subsetFile = system.file("settings", "phenotypeLibraryIds.csv", package = "rewardb"),
  removeExisting = FALSE,
  generateSql = TRUE
)
}
\arguments{
\item{connection}{DatabaseConnector::connection to postgres}

\item{config}{Global rewardb config object}

\item{libraryRepo}{Path to the repository. If local = TRUE path to custom git should be set here. Otherwise must specify a github repo}

\item{ref}{git ref to use. Defaults to master. could be a specific git tag or branch}

\item{local}{local copy or github}

\item{packageName}{name of package to install}

\item{subsetFile}{string path to subset}

\item{removeExisting}{remove existing pacakge of the same name}

\item{generateSql}{use web api to generate sql from source}
}
\description{
Add the OHDSI phenotype library references to the db.
This method heavily depends on the format of the Phenotype Library staying solid which is unlikely at this stage.
By default, uses remotes::install_github but can be set to a custom, local git repository
}
