% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyBoxPlotModule.R
\name{boxPlotModuleServer}
\alias{boxPlotModuleServer}
\title{Returns a reference to a server function}
\usage{
boxPlotModuleServer(distStatsFunc, caption, selectedExposureOutcome)
}
\arguments{
\item{distStatsFunc}{call to model that takes treament id and outcome id and returns stats dataframe}

\item{caption}{text caption string}

\item{selectedExposureOutcome}{exposure outcome reactive}
}
\description{
Returns a reference to a server function
}
