% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildDashboard.R
\name{buildDashboardFromConfig}
\alias{buildDashboardFromConfig}
\title{Build dashboard from config file}
\usage{
buildDashboardFromConfig(
  filePath,
  globalConfigPath,
  performCalibration = TRUE,
  allowUserAccess = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{filePath}{- path to a yaml configuration file used}

\item{globalConfigPath}{path to global reward config containing postgres db connection details}

\item{performCalibration}{- use empirical calibration package to compute adjusted p values, effect estimates and confidence intervals}

\item{allowUserAccess}{- enables grant permission for read only database user}
}
\description{
Builds a rewardb dashboard for all exposures for subset outcomes or all outcomes for subset exposures
Requires the rewardb results to be generated already.
This exports the data to a db backend and allows the config file to be used to run the shiny dashboard
}
