% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildPgDatabase.R
\name{buildPgDatabase}
\alias{buildPgDatabase}
\title{Build postgres schema}
\usage{
buildPgDatabase(
  configFilePath = "config/global-cfg.yml",
  buildPhenotypeLibrary = TRUE,
  generatePlSql = TRUE
)
}
\arguments{
\item{configFilePath}{path to global reward config}

\item{buildPhenotypeLibrary}{optionally add the entire phenotype library github R package}

\item{generatePlSql}{If building the phenotype library, generate the SQL from the github defintion or not (if false this assumes the git defintion works)}
}
\description{
Build the reward database schema for postgres instance from scratch
This will also require a CEM schema to be built which uses the OHDSI Common Evidence Model to generate the matrix
of known assocations for OMOP Standard Vocabulary terms. This is required for generating any stats that require negative controls
}
