% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calibration.R
\name{calibrateCustomCohorts}
\alias{calibrateCustomCohorts}
\title{Compute the calibrated results for custom cohort outcomes
Requires negative control cohorts to be set}
\usage{
calibrateCustomCohorts(appContext)
}
\arguments{
\item{appContext}{takes a rewardb application context}

\item{targetCohortIds}{- these are the cohort ids in the db and (currently) must be specified manually. TODO: remove}
}
\description{
Compute the calibrated results for custom cohort outcomes
Requires negative control cohorts to be set
}
