% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calibration.R
\name{calibrateOutcomes}
\alias{calibrateOutcomes}
\title{Compute the calibrated results for cohort outcomes
Requires negative control cohorts to be set}
\usage{
calibrateOutcomes(appContext)
}
\arguments{
\item{appContext}{takes a rewardb application context}
}
\description{
Compute the calibrated results for cohort outcomes
Requires negative control cohorts to be set
}
