% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calibration.R
\name{calibrateOutcomesCustomCohorts}
\alias{calibrateOutcomesCustomCohorts}
\title{Compute the calibrated results for custom cohort outcomes
Requires negative control cohorts to be set}
\usage{
calibrateOutcomesCustomCohorts(appContext)
}
\arguments{
\item{appContext}{takes a rewardb application context}
}
\description{
Compute the calibrated results for custom cohort outcomes
Requires negative control cohorts to be set
}
