% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildCohorts.R
\name{computeAtlasCohorts}
\alias{computeAtlasCohorts}
\title{Compute Atlas cohorts}
\usage{
computeAtlasCohorts(connection, config, exposureCohorts = FALSE)
}
\arguments{
\item{connection}{DatabaseConnector connection to cdm}

\item{config}{cdmConfig}

\item{exposureCohorts}{get exposures or not}
}
\description{
Computes sql cohorts against the CDM
}
