% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{computeCalibratedRows}
\alias{computeCalibratedRows}
\title{Actual calibration is performed here in a dplyr friendly way}
\usage{
computeCalibratedRows(interest, negatives, idCol, calibrationType = 1)
}
\arguments{
\item{interest}{this is the cohort set that should be calibrated}

\item{negatives}{these are the negative control cohort results}

\item{idCol}{- either target_cohort_id or outcome_cohort_id, this function is used in p}
}
\value{
data.frame
}
\description{
Actual calibration is performed here in a dplyr friendly way
}
