% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreComputeNulls.R
\name{computeExposureNullDistributions}
\alias{computeExposureNullDistributions}
\title{Compute null exposure distributions}
\usage{
computeExposureNullDistributions(
  config,
  analysisId = 1,
  sourceIds = NULL,
  nThreads = 10,
  minCohortSize = 5
)
}
\description{
Compute the null exposure distribution for all exposure cohorts (requires results and cem_matrix summary table)
}
