% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildDashboard.R
\name{computeMetaAnalysis}
\alias{computeMetaAnalysis}
\title{compute meta analysis}
\usage{
computeMetaAnalysis(appContext, connection)
}
\arguments{
\item{appContext}{application context loaded from yaml}

\item{connection}{DatabaseConnector connection to cdm}
}
\description{
Runs and saves metanalayis on data, uploads back to db
}
