% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreComputeNulls.R
\name{computeOutcomeNullDistributions}
\alias{computeOutcomeNullDistributions}
\title{Compute null exposure distributions}
\usage{
computeOutcomeNullDistributions(config, analysisId = 1, nThreads = 10)
}
\description{
Compute the null exposure distribution for all exposure cohorts (requires results and cem_matrix summary table)
}
