% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildReferenceTables.R
\name{createAtlasReference}
\alias{createAtlasReference}
\title{Create atlas cohort references in specified data source}
\usage{
createAtlasReference(connection, config, dataSource, customOutcomeCohortList)
}
\arguments{
\item{connection}{DatabaseConnector connection to cdm}

\item{customOutcomeCohortList}{custom cohorts to create}

\item{dataSources}{dataSource to run cohort on}
}
\description{
Create atlas cohort references in specified data source
}
