% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildCohorts.R
\name{createCohorts}
\alias{createCohorts}
\title{Create exposure cohorts}
\usage{
createCohorts(connection, config, deleteExisting = FALSE)
}
\arguments{
\item{connection}{DatabaseConnector connection to cdm}

\item{config}{cdm config}

\item{deleteExisting}{delete any existing computed cohorts}
}
\description{
Create all the exposure cohorts on a CDM
Note, will not recompute if they already exist
}
