% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildCohorts.R
\name{createCustomDrugEras}
\alias{createCustomDrugEras}
\title{create the custom drug eras, these are for drugs with nonstandard eras (e.g. where doeses aren't picked up by
repeat perscriptions). Could be something like a vaccine where exposed time is non trivial.}
\usage{
createCustomDrugEras(connection, config, dataSources)
}
\arguments{
\item{connection}{DatabaseConnector connection to cdm}
}
\description{
create the custom drug eras, these are for drugs with nonstandard eras (e.g. where doeses aren't picked up by
repeat perscriptions). Could be something like a vaccine where exposed time is non trivial.
}
