% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildDashboard.R
\name{createDashSchema}
\alias{createDashSchema}
\title{Create Dashboard Schema}
\usage{
createDashSchema(appContext, connection)
}
\arguments{
\item{appContext}{application context loaded from yaml}

\item{connection}{DatabaseConnector connection}
}
\description{
Creates postgres db dashboard schema for study
}
