% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildDashboardDatabase.R
\name{createOutcomeConceptMapping}
\alias{createOutcomeConceptMapping}
\title{Map outcomes to concept ids}
\usage{
createOutcomeConceptMapping(appContext)
}
\arguments{
\item{appContext}{rewardb app context}
}
\description{
Map outcomes to concept ids
}
