% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildReferenceTables.R
\name{createReferenceTables}
\alias{createReferenceTables}
\title{Create reference tables used to compute things in rewardb
TODO: split up data sources and main reference tables in to different function calls}
\usage{
createReferenceTables(connection, config, dataSources)
}
\arguments{
\item{connection}{DatabaseConnector connection to cdm}

\item{dataSources}{dataSources to run cohort on}
}
\description{
Create reference tables used to compute things in rewardb
TODO: split up data sources and main reference tables in to different function calls
}
