% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateOutcomeSummary.R
\name{createSummaryTables}
\alias{createSummaryTables}
\title{Create outcome summary tables}
\usage{
createSummaryTables(connection, config, dataSources)
}
\arguments{
\item{connection}{DatabaseConnector connection object}
}
\description{
Create outcome summary tables
}
