% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildReferenceTables.R
\name{createTargetDefinitions}
\alias{createTargetDefinitions}
\title{Names for target cohorts}
\usage{
createTargetDefinitions(connection, config)
}
\arguments{
\item{connection}{DatabaseConnector connection to cdm}
}
\description{
Names for target cohorts
}
