% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyServer.R
\name{dashboardInstance}
\alias{dashboardInstance}
\title{Dashboard instance}
\usage{
dashboardInstance(input, output, session)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session}
}
\description{
Requires a server appContext instance to be loaded in environment see scoping of launchDashboard
This can be obtained with rewardb::loadAppContext(...)
}
