% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyDeploy.R
\name{deployShinyApp}
\alias{deployShinyApp}
\title{Deploy shiny application}
\usage{
deployShinyApp(
  appConfigPath,
  user,
  globalConfigPath = "~/reward/config/global-cfg.yml",
  shinyServer = "sharedshiny-prod.jnj.com",
  port = 22,
  deployPath = "~/ShinyApps/reward/",
  renvPath = "~/reward/",
  keyfile = NULL
)
}
\description{
Deploy shiny dashboards to shiny server via ssh
}
