% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExecuteDataBuild.R
\name{.getLogger}
\alias{.getLogger}
\title{get logger}
\usage{
.getLogger(logFileName, .clearLoggers = TRUE)
}
\arguments{
\item{logFileName}{path to log file}

\item{.clearLoggers}{clear existing loggers and make new one}
}
\description{
Get the logger and set where the log file is stored.
}
