% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReferenceIO.R
\name{exportAtlasCohortRef}
\alias{exportAtlasCohortRef}
\title{exportAtlasCohortRef}
\usage{
exportAtlasCohortRef(
  config,
  atlasIds,
  exportZipFile,
  atlasUrl = NULL,
  exportPath = tempdir()
)
}
\arguments{
\item{config}{rewardb global config}

\item{atlasIds}{ids to atlas cohort to pull down}

\item{connection}{DatabaseConnector::connection to postgres}
}
\description{
Exports atlas cohort references to a zip file for import in to an already existing db
}
