% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsIO.R
\name{exportResults}
\alias{exportResults}
\title{Generic function that creates a zip file of csvs that will be imported in to the reward postgres instance}
\usage{
exportResults(
  config,
  exportZipFile = "rewardb-export.zip",
  csvPattern = "*.csv",
  tableNames = list(),
  cdmVersion = NULL,
  databaseId = NULL,
  exportPath = NULL
)
}
\description{
Generic function that creates a zip file of csvs that will be imported in to the reward postgres instance
}
