% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildDashboardDatabase.R
\name{extractOutcomeCohortNames}
\alias{extractOutcomeCohortNames}
\title{Map outcome cohorts to names}
\usage{
extractOutcomeCohortNames(appContext)
}
\arguments{
\item{appContext}{rewardb app context}
}
\description{
Map outcome cohorts to names
}
