% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildDashboard.R
\name{extractResultsSubset}
\alias{extractResultsSubset}
\title{Takes results from a CDM and puts them in the rewardb schema}
\usage{
extractResultsSubset(appContext)
}
\arguments{
\item{appContext}{rewardb app context}
}
\description{
Takes results from a CDM and puts them in the rewardb schema
}
