% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildDashboard.R
\name{extractTargetCohortNames}
\alias{extractTargetCohortNames}
\title{Map target cohorts to names and exposure classes}
\usage{
extractTargetCohortNames(appContext)
}
\arguments{
\item{appContext}{rewardb app context}
}
\description{
Map target cohorts to names and exposure classes
}
