% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExecuteDataBuild.R
\name{fullExecution}
\alias{fullExecution}
\title{Run a full execution of the rewardb pipeline on the CDM sets specified
This creates a set of reference tables, creates exposure cohorts, creates outcome cohorts, generates summary tables
of combined outcome, exposure pairs and runs SCC analysis.
if dataSources is set to a vector then the anlaysis will only be performed on this subset - these must be specified in the config yml file}
\usage{
fullExecution(
  configFilePath = "config/global-cfg.yml",
  .createReferences = TRUE,
  .addDefaultAtlasCohorts = TRUE,
  .createExposureCohorts = TRUE,
  .createOutcomeCohorts = TRUE,
  .generateSummaryTables = TRUE,
  .runSCC = TRUE,
  dataSources = NULL,
  logFileName = "rbDataBuild.log"
)
}
\arguments{
\item{configFilePath}{path on disk to a yaml configuration file containing all needed config setting to point at relevant cdm}

\item{.createReferences}{create the reference tables (set to false if already ran)}

\item{.addDefaultAtlasCohorts}{use default cohort ids speciefied in config file (these can be added later manually)}

\item{.createExposureCohorts}{Create the drug exposure cohorts default is TRUE}

\item{.createOutcomeCohorts}{Create the outcome cohorts, default is TRUE}

\item{.generateSummaryTables}{Generates summaries for outcome, target pairs}

\item{.runSCC}{perform self controlled cohort analysis and add results to the merged output db}

\item{dataSources}{vector of strings or null - keys to cdm stores to use. By default all cdms are added for a given config file}
}
\description{
Run a full execution of the rewardb pipeline on the CDM sets specified
This creates a set of reference tables, creates exposure cohorts, creates outcome cohorts, generates summary tables
of combined outcome, exposure pairs and runs SCC analysis.
if dataSources is set to a vector then the anlaysis will only be performed on this subset - these must be specified in the config yml file
}
