% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortStats.R
\name{getAverageTimeOnTreatment}
\alias{getAverageTimeOnTreatment}
\title{Get Self-Controlled cohort stats}
\usage{
getAverageTimeOnTreatment(
  config,
  analysisOptions = list(),
  analysisId = NULL,
  targetCohortIds = NULL,
  outcomeCohortIds = NULL
)
}
\arguments{
\item{config}{reward cdm config object}

\item{analysisOptions}{arguments for SCC}

\item{analysisId}{Id for storage in returned datafame}

\item{targetCohortIds}{target cohort ids}

\item{outcomeCohortIds}{outcome cohort ids}
}
\value{
data.frame of stats mapped for reward db
}
\description{
compute average time on treatement and average time to outcome
}
