% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calibration.R
\name{getCalibratedAtlasTargets}
\alias{getCalibratedAtlasTargets}
\title{Get calibrated Atlas cohorts}
\usage{
getCalibratedAtlasTargets(appContext, connection)
}
\arguments{
\item{appContext}{rewardb app context}
}
\value{
data.frame of uncalibrated results
}
\description{
Compute calibtation on atlas targets
}
