% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calibration.R
\name{getCalibratedGenericTargets}
\alias{getCalibratedGenericTargets}
\title{Get calibrated generic, not atlas cohorts}
\usage{
getCalibratedGenericTargets(appContext, connection)
}
\arguments{
\item{appContext}{rewardb app context}
}
\value{
data.frame of calibrated results
}
\description{
Compute calibtation on atlas targets
}
