% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calibration.R
\name{getCalibratedOutcomes}
\alias{getCalibratedOutcomes}
\title{Get all calibrated outcome cohorts}
\usage{
getCalibratedOutcomes(appContext, connection)
}
\arguments{
\item{appContext}{takes a rewardb application context}
}
\value{
data.frame of calibrated results
}
\description{
Compute the calibrated results for cohort outcomes
Requires negative control cohorts to be set
Used automatically when appContext$useExposureControls is TRUE
}
