% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calibration.R
\name{getCalibratedTargets}
\alias{getCalibratedTargets}
\title{Get all calibrated target cohorts}
\usage{
getCalibratedTargets(appContext, connection)
}
\arguments{
\item{appContext}{takes a rewardb application context}
}
\value{
data.frame of calibrated results
}
\description{
Compute the calibrated results for cohort targets
Requires negative control cohorts to be set
}
