% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreComputeNulls.R
\name{getCemExposureNegativeControlConcepts}
\alias{getCemExposureNegativeControlConcepts}
\title{Gets negative control outcome concepts for exposure cohorts}
\usage{
getCemExposureNegativeControlConcepts(globalConfig)
}
\description{
Gets negative control outcome concepts for exposure cohorts
}
