% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreComputeNulls.R
\name{getCemOutcomeNegativeControlConcepts}
\alias{getCemOutcomeNegativeControlConcepts}
\title{Gets negative control outcome concepts for exposure cohorts}
\usage{
getCemOutcomeNegativeControlConcepts(globalConfig)
}
\description{
Gets negative control outcome concepts for exposure cohorts
}
