% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{getConcetptCohortData}
\alias{getConcetptCohortData}
\title{From pairs of exposure, outcomes get dataframe of assocaited RR values from reward data}
\usage{
getConcetptCohortData(
  connection,
  config,
  outcomeExposurePairs,
  snakeCaseToCamelCase = TRUE,
  outcomeTypes = 0
)
}
\arguments{
\item{connection}{DatabaseConnector connection object}

\item{config}{Reward global config object}

\item{outcomeExposurePairs}{Dataframe containing the coulmns 'exposureConceptId' and 'outcomeConceptId'}

\item{snakeCaseToCamelCase}{convert snake case to camle case in output frame}
}
\value{
Dataframe containing RR results for cohorts that match the concept id pairs
}
\description{
Function to get cohorts from outcome/exposure pairs. Will get any cohorts results for cohorts
that match these pairs including any ancestor concepts.
}
