% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExecuteDataBuild.R
\name{getDatabaseId}
\alias{getDatabaseId}
\title{getCdmVersion}
\usage{
getDatabaseId(cdmConfig)
}
\arguments{
\item{cdmConfig}{reward cdmConfig loaded with cdmConfig}

\item{connection}{DatabaseConnector connection}
}
\description{
Get the database version if a _version table exists. This is a JNJ standard but not a requirment of a cdm
Returns -1 where table is not present. Only really used for debugging in meta-data
}
