% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evidence.R
\name{getMappedControls}
\alias{getMappedControls}
\title{Get mapped controlls}
\usage{
getMappedControls(
  connection,
  targetCohortIds,
  outcomeCohortIds,
  schema,
  cemSchema,
  vocabularySchema,
  summaryTable = "matrix_summary"
)
}
\arguments{
\item{connection}{DatabaseConnector connection}

\item{targetCohortIds}{condition cohort}

\item{outcomeCohortIds}{drug cohort ids}

\item{schema}{dashboard schema}

\item{cemSchema}{CEM schema}

\item{vocabularySchema}{vocabulary schema}

\item{summaryTable}{schma for summary of CEM results - probably matrix summary}
}
\value{
data.frame of negative controlls mapped by concept
}
\description{
Gets mapped concepts from cem schema
}
