% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{getMetaAnalysisData}
\alias{getMetaAnalysisData}
\title{Get Meta Analysis Data}
\usage{
getMetaAnalysisData(table)
}
\arguments{
\item{table}{expected data.frame containing fielsds: T_AT_RISK, T_PT, T_CASES, C_AT_RISK, C_PT, C_CASES, IRR}
}
\description{
Peform meta analysis on a dataframe of elements and return row (e.g. to be appended with rbind
}
