% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calibration.R
\name{getOutcomeControls}
\alias{getOutcomeControls}
\title{Get outcome controls}
\usage{
getOutcomeControls(
  appContext,
  connection,
  targetIds = NULL,
  outcomeTypes = c(0, 1, 2),
  minCohortSize = 5
)
}
\arguments{
\item{appContext}{rewardb app context}

\item{connection}{DatabaseConnector connection to postgres rewardb instance}

\item{targetIds}{cohorts to get negative controls for}

\item{minCohortSize}{smaller cohorts are not generally used for calibration as rr values tend to be extremes}
}
\value{
data.frame of negative control exposures for specified outcomes
}
\description{
Returns a set of specified negative controls for input target cohorts
}
