% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Config.R
\name{getPasswordSecurely}
\alias{getPasswordSecurely}
\title{Get password securely}
\usage{
getPasswordSecurely(
  envVar = "REWARD_PASSWORD",
  prompt = "Enter the reward database password"
)
}
\arguments{
\item{envVar}{environment variable to store password in}

\item{promt}{text prompt when loading askpass prompt}
}
\value{
string from prompt or environment variable
}
\description{
Gets password from user. Ignored if REWARD_B_PASSWORD system env variable is set (e.g. in .Rprofile)
}
