% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExecuteDataBuild.R
\name{getSccResults}
\alias{getSccResults}
\title{Get Zipped Scc Results}
\usage{
getSccResults(
  config,
  connection,
  configId,
  outcomeCohortIds = NULL,
  targetCohortIds = NULL
)
}
\arguments{
\item{config}{- cdm config loaded with loadCdmConfig function}

\item{connection}{DatabaseConnector connection}

\item{outcomeCohortIds}{- vector of outcome cohort ids or NULL}

\item{targetCohortIds}{- vector of exposure cohort ids or NULL}

\item{.generateCohortStats}{- generate time on treatment and time to outcome stats or not}
}
\value{
list of zip file locations
}
\description{
Get zip files for scc
Partial reward execution with a subset of targets or outcomes. If both are null this will generate SCC results for all
exposure and outcome pairs. This is only really useful if you're adding an cohort after the full result set has been
generated.
}
