% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortStats.R
\name{getSelfControlledCohortExposureStats}
\alias{getSelfControlledCohortExposureStats}
\title{Get Self-Controlled cohort stats}
\usage{
getSelfControlledCohortExposureStats(
  connectionDetails,
  cdmDatabaseSchema,
  exposureIds,
  outcomeIds,
  outcomeDatabaseSchema,
  exposureDatabaseSchema,
  outcomeTable = "condition_era",
  exposureTable = "drug_era",
  oracleTempSchema = NULL,
  firstExposureOnly = TRUE,
  firstOutcomeOnly = TRUE,
  minAge = "",
  maxAge = "",
  studyStartDate = "",
  studyEndDate = "",
  addLengthOfExposureExposed = TRUE,
  riskWindowStartExposed = 1,
  riskWindowEndExposed = 1,
  addLengthOfExposureUnexposed = TRUE,
  riskWindowEndUnexposed = -1,
  riskWindowStartUnexposed = -1,
  hasFullTimeAtRisk = TRUE,
  washoutPeriod = 0,
  followupPeriod = 0
)
}
\arguments{
\item{connectionDetails}{An R object of type \code{connectionDetails} created using
the function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{cdmDatabaseSchema}{Name of database schema that contains the OMOP CDM and
vocabulary.}

\item{exposureIds}{A vector containing the drug_concept_ids or
cohort_definition_ids of the exposures of interest. If empty,
all exposures in the exposure table will be included.}

\item{outcomeIds}{The condition_concept_ids or cohort_definition_ids of the
outcomes of interest. If empty, all the outcomes in the
outcome table will be included.}

\item{outcomeDatabaseSchema}{The name of the database schema that is the location where
the data used to define the outcome cohorts is available. If
exposureTable = CONDITION_ERA, exposureDatabaseSchema is not
used by assumed to be cdmSchema.  Requires read permissions
to this database.}

\item{exposureDatabaseSchema}{The name of the database schema that is the location where
the exposure data used to define the exposure cohorts is
available. If exposureTable = DRUG_ERA,
exposureDatabaseSchema is not used by assumed to be
cdmSchema.  Requires read permissions to this database.}

\item{outcomeTable}{The tablename that contains the outcome cohorts.  If
outcomeTable <> CONDITION_OCCURRENCE, then expectation is
outcomeTable has format of COHORT table:
COHORT_DEFINITION_ID, SUBJECT_ID, COHORT_START_DATE,
COHORT_END_DATE.}

\item{exposureTable}{The tablename that contains the exposure cohorts.  If
exposureTable <> DRUG_ERA, then expectation is exposureTable
has format of COHORT table: cohort_concept_id, SUBJECT_ID,
COHORT_START_DATE, COHORT_END_DATE.}

\item{oracleTempSchema}{For Oracle only: the name of the database schema where you
want all temporary tables to be managed. Requires
create/insert permissions to this database.}

\item{firstExposureOnly}{If TRUE, only use first occurrence of each drug concept idgetSccStats
for each person}

\item{firstOutcomeOnly}{If TRUE, only use first occurrence of each condition concept
id for each person.}

\item{minAge}{Integer for minimum allowable age.}

\item{maxAge}{Integer for maximum allowable age.}

\item{studyStartDate}{Date for minimum allowable data for index exposure. Date
format is 'yyyymmdd'.}

\item{studyEndDate}{Date for maximum allowable data for index exposure. Date
format is 'yyyymmdd'.}

\item{addLengthOfExposureExposed}{If TRUE, use the duration from drugEraStart -> drugEraEnd as
part of timeAtRisk.}

\item{riskWindowStartExposed}{Integer of days to add to drugEraStart for start of
timeAtRisk (0 to include index date, 1 to start the day
after).}

\item{riskWindowEndExposed}{Additional window to add to end of exposure period (if
addLengthOfExposureExposed = TRUE, then add to exposure end
date, else add to exposure start date).}

\item{addLengthOfExposureUnexposed}{If TRUE, use the duration from exposure start -> exposure
end as part of timeAtRisk looking back before exposure
start.}

\item{riskWindowEndUnexposed}{Integer of days to add to exposure start for end of
timeAtRisk (0 to include index date, -1 to end the day
before).}

\item{riskWindowStartUnexposed}{Additional window to add to start of exposure period (if
addLengthOfExposureUnexposed = TRUE, then add to exposure
end date, else add to exposure start date).}

\item{hasFullTimeAtRisk}{If TRUE, restrict to people who have full time-at-risk
exposed and unexposed.}

\item{washoutPeriod}{Integer to define required time observed before exposure
start.}

\item{followupPeriod}{Integer to define required time observed after exposure
start.}

\item{cdmVersion}{Define the OMOP CDM version used: currently support "4" and
"5".}

\item{computeThreads}{Number of parallel threads for computing IRRs with exact
confidence intervals.}
}
\value{
A data frame  containing the results of the analysis.
}
\description{
Generate time to outcome and time on treatment statistics.
Takes the same parameters as the self-controlled cohort call
}
\examples{
\dontrun{
connectionDetails <- createConnectionDetails(dbms = "sql server",
                                             server = "RNDUSRDHIT07.jnj.com")
sccResult <- getSelfControlledCohortExposureStats(connectionDetails,
                                     cdmDatabaseSchema = "cdm_truven_mdcr.dbo",
                                     exposureIds = c(767410, 1314924, 907879),
                                     outcomeIds = 444382,
                                     outcomeTable = "condition_era")
}
}
