% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evidence.R
\name{getStudyControls}
\alias{getStudyControls}
\title{Get mapped controlls}
\usage{
getStudyControls(
  connection,
  schema,
  cemSchema,
  vocabularySchema,
  targetCohortIds = NULL,
  outcomeCohortIds = NULL
)
}
\arguments{
\item{connection}{DatabaseConnector connection}

\item{schema}{dashboard schema}

\item{cemSchema}{CEM schema}

\item{vocabularySchema}{vocabulary schema}

\item{targetCohortIds}{condition concept ids}

\item{outcomeCohortIds}{drug concept ids}
}
\value{
data.frame of negative controlls mapped by concept
}
\description{
From  outcome and target cohort ids, construct tables that contain their concept ids
}
