% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Config.R
\name{getTargetCohortIds}
\alias{getTargetCohortIds}
\title{getTargetCohortIds}
\usage{
getTargetCohortIds(appContext, connection)
}
\arguments{
\item{appContext}{application context}

\item{defaults}{list of default values to check and set if null}
}
\value{
updated list
}
\description{
Get cohorts used on dashboard
}
