% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calibration.R
\name{getUncalibratedAtlasCohorts}
\alias{getUncalibratedAtlasCohorts}
\title{Get uncalibrated Atlas cohorts}
\usage{
getUncalibratedAtlasCohorts(appContext)
}
\arguments{
\item{appContext}{rewardb app context}
}
\value{
data.frame of uncalibrated results from databases
}
\description{
Get atlas cohorts results pre-calibration
}
