% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{getUncalibratedOutcomes}
\alias{getUncalibratedOutcomes}
\title{get outcomes that are not calibrated
param appContext rewardb app context}
\usage{
getUncalibratedOutcomes(appContext, targetCohortIds)
}
\value{
data.frame of uncalibrated results from databases
}
\description{
get outcomes that are not calibrated
param appContext rewardb app context
}
