% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildCohorts.R
\name{getUncomputedAtlasCohorts}
\alias{getUncomputedAtlasCohorts}
\title{get Uncomputed Atlas Cohorts}
\usage{
getUncomputedAtlasCohorts(connection, config, exposureCohorts = FALSE)
}
\arguments{
\item{connection}{connection}

\item{config}{cdmConfig}

\item{exposureCohorts}{get exposures or not}
}
\description{
Get cohorts that haven't been computed and return their references from file on disk
SQL and JSON references are not stored in the CDM database's scratch schema
}
