% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Atlas.R
\name{getWebObject}
\alias{getWebObject}
\title{Hit a web api REST endpoint
TODO: replace with ROhdsiWebApi calls}
\usage{
getWebObject(webApiUrl, resource, id)
}
\arguments{
\item{webApiUrl}{base webApi url}

\item{resource}{url to hit}

\item{id}{url id (e.g. atlas cohort id)}
}
\description{
Hit a web api REST endpoint
TODO: replace with ROhdsiWebApi calls
}
