% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReferenceIO.R
\name{importReferenceTables}
\alias{importReferenceTables}
\title{Import reference tables}
\usage{
importReferenceTables(cdmConfig, zipFilePath, usePgCopy = FALSE)
}
\arguments{
\item{cdmConfig}{cdmConfig object}

\item{zipFilePath}{zip file path}

\item{usePgCopy}{Use postgre copy function (not included in database connector)}
}
\description{
Note that this always overwrites the existing reference tables stored in the database
}
