% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsIO.R
\name{importResultsFiles}
\alias{importResultsFiles}
\title{Only works with postgres}
\usage{
importResultsFiles(
  connectionDetails,
  resultsSchema,
  exportZipFilePath,
  unzipPath = "rb-import",
  .checkTables = FALSE,
  .debug = FALSE
)
}
\description{
Only works with postgres
}
