% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Atlas.R
\name{insertCustomExposureRef}
\alias{insertCustomExposureRef}
\title{Insert atlas concept set to postgres db for a custom exposure cohort}
\usage{
insertCustomExposureRef(
  connection,
  config,
  conceptSetId,
  cohortName,
  conceptSetDefinition = NULL,
  .warnNonIngredients = TRUE,
  webApiUrl = NULL
)
}
\arguments{
\item{connection}{DatabaseConnector::connection to cdm}

\item{config}{rewardb global config}

\item{conceptSetId}{id to atlas cohort to pull down}

\item{cohortName}{Name to give cohort}

\item{conceptSetDefinition}{webApi response object}

\item{.warnNonIngredients}{Warn for non ingredients}

\item{webApiUrl}{URL of webAPI service (if different from specified in config)}
}
\description{
Adds atlas concept set to db reference, from web api
Inserts name/id in to custom cohort table
Maps condition concepts of interest, any desecdants or if they're excluded from the cohort
deprecated
Use expsosure = TRUE with atlas cohort definitions and add them will be removed in future version
}
