% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyServer.R
\name{launchDashboard}
\alias{launchDashboard}
\title{Launch the REWARD Shiny app dashboard}
\usage{
launchDashboard(appConfigPath, globalConfigPath)
}
\arguments{
\item{appConfigPath}{path to configuration file. This is loaded in to the local environment with the appContext variable}
}
\description{
Launch the REWARD Shiny app dashboard
}
\details{
Launches a Shiny app for a given configuration file
}
