% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Explorer.R
\name{launchExplorer}
\alias{launchExplorer}
\title{Launch the REWARD Shiny app report}
\usage{
launchExplorer(globalConfigPath)
}
\arguments{
\item{appConfigPath}{path to configuration file. This is loaded in to the local environment with the appContext variable}
}
\description{
Launches a Shiny app for a given configuration file
}
