% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyServer.R
\name{launchReport}
\alias{launchReport}
\title{Launch the REWARD Shiny app report}
\usage{
launchReport(globalConfigPath, exposureId = NULL, outcomeId = NULL)
}
\arguments{
\item{exposureId}{exposure cohort id}

\item{outcomeId}{outcome cohort id}

\item{appConfigPath}{path to configuration file. This is loaded in to the local environment with the appContext variable}
}
\description{
Launches a Shiny app for a given configuration file
}
