% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readConfig.R
\name{loadAppContext}
\alias{loadAppContext}
\title{Loads the application configuration and creates an application object
By default, loads the database connections in to this object
loads database password from prompt if REWARD_B_PASSWORD system env variable is not set (e.g. in .Rprofile)
The idea is to allow shared configuration settings between the web app and any data processing tools}
\usage{
loadAppContext(
  configPath,
  createConnection = FALSE,
  useCdm = FALSE,
  .env = .GlobalEnv
)
}
\arguments{
\item{configPath}{is a yaml file for the application configuration}

\item{create}{the database connections for this app - defaults to true}
}
\description{
Loads the application configuration and creates an application object
By default, loads the database connections in to this object
loads database password from prompt if REWARD_B_PASSWORD system env variable is not set (e.g. in .Rprofile)
The idea is to allow shared configuration settings between the web app and any data processing tools
}
\examples{
loadAppContext('config/config.dev.yml')
}
\keyword{appContext}
