% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Config.R
\name{loadCdmConfig}
\alias{loadCdmConfig}
\title{load cdm config object}
\usage{
loadCdmConfig(cdmConfigPath)
}
\arguments{
\item{cdmConfigPath}{cdmConfigPath}
}
\description{
Loads config and prompt user for db password
Password can be set in envrionment variable passwordEnvironmentVariable of yaml file
}
