% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conf.R
\name{loadCdmConfiguration}
\alias{loadCdmConfiguration}
\title{load cdm config object}
\usage{
loadCdmConfiguration(cdmConfigPath)
}
\arguments{
\item{cdmConfigPath}{cdmConfigPath}
}
\description{
Loads config and prompt user for db password
Password can be set in envrionment variable passwordEnvironmentVariable of yaml file
}
