% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conf.R
\name{loadGlobalConfiguration}
\alias{loadGlobalConfiguration}
\title{Loads global config}
\usage{
loadGlobalConfiguration(globalConfigPath)
}
\arguments{
\item{globalConfigPath}{path to global yaml}
}
\description{
Load reward global config yaml file
}
