% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{loadRenderTranslateExecuteSql}
\alias{loadRenderTranslateExecuteSql}
\title{Load, render, translate and execute a SQL file in this package.}
\usage{
loadRenderTranslateExecuteSql(
  connection,
  sqlFilename,
  dbms = "sql server",
  ...,
  snakeCaseToCamelCase = FALSE,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  warnOnMissingParameters = TRUE
)
}
\arguments{
\item{sqlFilename}{The source SQL file}

\item{dbms}{The target dialect. Currently 'sql server', 'oracle', 'postgres',
and 'redshift' are supported}

\item{...}{Parameter values used for \code{render}}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support
temp tables. To emulate temp tables, provide a schema with write
privileges where temp tables can be created.}

\item{warnOnMissingParameters}{Should a warning be raised when parameters provided to this
function do not appear in the parameterized SQL that is being
rendered? By default, this is TRUE.}
}
\value{
Returns a string containing the rendered SQL.
}
\description{
NOTE: This function does not support dialect-specific SQL translation
at this time.
}
