% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Config.R
\name{loadReportContext}
\alias{loadReportContext}
\title{Load report application context}
\usage{
loadReportContext(
  globalConfigPath,
  .env = .GlobalEnv,
  exposureId = NULL,
  outcomeId = NULL
)
}
\arguments{
\item{globalConfigPath}{is a yaml file for the application configuratione}

\item{.env}{environment to load variable in to}

\item{exposureId}{exposure cohort id}

\item{outcomeId}{outcome cohort id}
}
\description{
By default, loads the database connections in to this object
loads database password from prompt if REWARD_B_PASSWORD system env variable is not set (e.g. in .Rprofile)
The idea is to allow shared configuration settings between the web app and any data processing tools
}
\examples{
loadAppContext('config/config.dev.yml', 'config/global-cfg.yml')
}
\keyword{reportAppContext}
