% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conf.R
\name{loadShinyAppContext}
\alias{loadShinyAppContext}
\title{Loads Application Context}
\usage{
loadShinyAppContext(configPath, globalConfigPath)
}
\arguments{
\item{configPath}{is a yaml file for the application configuration}

\item{globalConfigPath}{path to global yaml}
}
\description{
By default, loads the database connections in to this object
loads database password from prompt if REWARD_B_PASSWORD system env variable is not set (e.g. in .Rprofile)
The idea is to allow shared configuration settings between the web app and any data processing tools
}
