% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildDashboard.R
\name{metaAnalysis}
\alias{metaAnalysis}
\title{meta analysis}
\usage{
metaAnalysis(table)
}
\arguments{
\item{table}{data.frame}
}
\value{
data.frame - results of meta analysis
}
\description{
compute meta-analysis across data sources provided in table
Perform meta-analysis on data sources
}
