% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExecuteDataBuild.R
\name{oneOffSccResults}
\alias{oneOffSccResults}
\title{oneOffSccResults}
\usage{
oneOffSccResults(
  cdmConfigFilePath,
  configId,
  outcomeCohortIds = NULL,
  targetCohortIds = NULL,
  .generateCohortStats = TRUE,
  logFileName = NULL
)
}
\arguments{
\item{cdmConfigFilePath}{- path to cdm config loaded with loadCdmConfig function}

\item{configId}{- string id that will be used as a prefix to store results files}

\item{outcomeCohortIds}{- vector of outcome cohort ids or NULL}

\item{targetCohortIds}{- vector of exposure cohort ids or NULL}

\item{.generateCohortStats}{- generate time on treatment and time to outcome stats or not}

\item{logFileName}{logfile used. If null is based on the passed configId}

\item{getDbId}{- assumes CDM version is stored in the cdm_source table}
}
\description{
Partial reward execution with a subset of targets or outcomes. If both are null this will generate SCC results for all
exposure and outcome pairs. This is only really useful if you're adding an cohort after the full result set has been
generated.
}
