% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReferenceIO.R
\name{registerCdm}
\alias{registerCdm}
\title{getCdmVersion}
\usage{
registerCdm(connection, globalConfig, cdmConfig)
}
\arguments{
\item{connection}{DatabaseConnector connection}

\item{globalConfig}{globak reward config}

\item{cdmConfig}{reward cdmConfig loaded with cdmConfig}
}
\description{
Register a cdm with central postgres db
}
