% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Atlas.R
\name{removeCustomExposureCohort}
\alias{removeCustomExposureCohort}
\title{Remove exposure conceptset cohort}
\usage{
removeCustomExposureCohort(connection, config, conceptSetId, webApiUrl = NULL)
}
\arguments{
\item{connection}{DatabaseConnector::connection to cdm}

\item{config}{rewardb global config}

\item{conceptSetId}{ids to remove from db}

\item{webApiUrl}{URL of WebApi resource (or other source if not imported via web api)}
}
\description{
Removes atlas entries from the cdm store
deprecated
Use expsosure = TRUE with atlas cohort definitions and add them will be removed in future version
}
