% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewardb.R
\docType{package}
\name{rewardb}
\alias{rewardb}
\title{rewardb: REal-World Assesement of DRug performance
To strengthen and expand our existing observational health data analytics system that proactively provides
real-world evidence through the population-level effect estimation to enable exploration of performance of existing medical products.}
\section{Configuration}{
NA
}

\section{Data Generation}{
NA
}

\section{Shiny Dashboard}{
NA
}

