% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExecuteDataBuild.R
\name{runDatabuildJob}
\alias{runDatabuildJob}
\title{run Databuild Job}
\usage{
runDatabuildJob(
  cdmConfigPath,
  globalConfigPath,
  rewardReferenceZipPath,
  name,
  workingDir = "."
)
}
\arguments{
\item{Charachter}{globalConfigPath path to rewarb global configuration}
}
\description{
Run analysis job script in rstudio session, this won't block the rstudio thread so other tasks can be completed
NOTE: that this copies the global env (and the values passed will modify .GlobalEnv variables of the same name)
}
