% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortStats.R
\name{runExposedOutcomeFetaureExtraction}
\alias{runExposedOutcomeFetaureExtraction}
\title{For patients exposed to outcomes, compute feature extraction stats}
\usage{
runExposedOutcomeFetaureExtraction(
  connectionDetails,
  exposureIds,
  outcomeIds,
  outcomeDatabaseSchema,
  exposureDatabaseSchema,
  cdmDatabaseSchema,
  outcomeTable = "condition_era",
  exposureTable = "drug_era",
  oracleTempSchema = NULL,
  firstExposureOnly = TRUE,
  firstOutcomeOnly = TRUE,
  minAge = "",
  maxAge = "",
  studyStartDate = "",
  studyEndDate = "",
  addLengthOfExposureExposed = TRUE,
  riskWindowStartExposed = 1,
  riskWindowEndExposed = 1,
  addLengthOfExposureUnexposed = TRUE,
  riskWindowEndUnexposed = -1,
  riskWindowStartUnexposed = -1,
  hasFullTimeAtRisk = TRUE,
  washoutPeriod = 0,
  followupPeriod = 0,
  covariateSettings = NULL
)
}
\description{
For patients exposed to outcomes, compute feature extraction stats
}
