% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreComputeNulls.R
\name{runPreComputeNullDistributions}
\alias{runPreComputeNullDistributions}
\title{Run pre compute null distributions}
\usage{
runPreComputeNullDistributions(
  globalConfigPath,
  analysisId = 1,
  sourceIds = NULL,
  nThreads = 10,
  getCemMappings = TRUE,
  minCohortSize = 5
)
}
\arguments{
\item{globalConfigPath}{Charachter path to global conifg yaml}

\item{analysisId}{Integer analysis settings ID (default is 1)}

\item{sourceIds}{Integer vector of integer sources to run on (default is NULL, all sources)}

\item{nThreads}{Integer number of threads}

\item{minCohortSize}{Minimum size of cohort to use as a negative control. Default is 5}
}
\description{
Run task to pre-compute null distributions for all outcomes and exposures.
This process requires a CemConnector connection, this may be slow if using cem.ohdsi.org
}
