% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Atlas.R
\name{sccAdHocCohorts}
\alias{sccAdHocCohorts}
\title{Generate scc results for a one off atlas cohort (added after a full results run)}
\usage{
sccAdHocCohorts(
  cdmConfigPath,
  globalConfigPath,
  configId,
  atlasIds,
  sourceUrl,
  exposure = FALSE,
  referenceZipFile = NULL
)
}
\arguments{
\item{configId}{String configuration id}

\item{exposure}{boolean - is this for exposure or outcome cohorts}

\item{referenceZipFile}{path to ouputed zipFile of references if they need to be updated}

\item{cdmConfig}{cdmConfig object loaded with loadCdmConfig}
}
\description{
Removes atlas entries from the cdm store
}
