% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Atlas.R
\name{sccOneOffAtlasCohort}
\alias{sccOneOffAtlasCohort}
\title{Generate scc results for a one off atlas cohort (added after a full results run)}
\usage{
sccOneOffAtlasCohort(cdmConfig, zipFilePath, configId)
}
\arguments{
\item{cdmConfig}{cdmConfig object loaded with loadCdmConfig}

\item{zipFilePath}{path to ouputed zipFile}

\item{configId}{String configuration id}
}
\description{
Removes atlas entries from the cdm store
}
