% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyServer.R
\name{serverInstance}
\alias{serverInstance}
\title{Requires a server appContext instance to be loaded in environment see scoping of launchDashboard
This can be obtained with rewardb::loadAppContext(...)}
\usage{
serverInstance(input, output, session)
}
\arguments{
\item{input}{shiny input object}

\item{output}{shiny output object}
}
\description{
Requires a server appContext instance to be loaded in environment see scoping of launchDashboard
This can be obtained with rewardb::loadAppContext(...)
}
