% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{unzipAndVerify}
\alias{unzipAndVerify}
\title{Unzip and verify results zip with meta-data json}
\usage{
unzipAndVerify(exportZipFilePath, unzipPath, overwrite)
}
\arguments{
\item{exportZipFilePath}{zip file to inflate}

\item{unzipPath}{path to create}

\item{overwrite}{overwrite any existing}
}
\description{
Used to unzip and check all files in a zip folder with meta data file containing md5 hashes at time of creation
Used by both results generation and reference files
}
