% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyDeploy.R
\name{updateShinyServer}
\alias{updateShinyServer}
\title{Update shiny server}
\usage{
updateShinyServer(
  user,
  shinyServer = "sharedshiny-prod.jnj.com",
  port = 22,
  renvPath = "~/reward/",
  keyfile = NULL
)
}
\description{
Updates code in the remote ren package
}
