% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildDashboard.R
\name{addCemEvidence}
\alias{addCemEvidence}
\title{ing a CEM source finds any evidence related to conditions and exposures
This is used for the automated construction of negative control sets and the indication labels
TODO: test and improve this logic and move to its own R file}
\usage{
addCemEvidence(appContext)
}
\arguments{
\item{appContext}{rewardb app context}
}
\description{
ing a CEM source finds any evidence related to conditions and exposures
This is used for the automated construction of negative control sets and the indication labels
TODO: test and improve this logic and move to its own R file
}
