% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildDashboardDatabase.R
\name{buildFromConfig}
\alias{buildFromConfig}
\title{Builds a rewardb dashboard for all exposures for subset outcomes or all outcomes for subset exposures
Requires the rewardb results to be generated already.
This exports the data to a db backend and allows the config file to be used to run the shiny dashboard}
\usage{
buildFromConfig(filePath, calibrateOutcomes = FALSE, calibrateTargets = FALSE)
}
\arguments{
\item{filePath}{- path to a yaml configuration file used}

\item{calibrateOutcomes}{- where subset of outcomes is specified}

\item{calibrateTargets}{- where a subset of target exposures is specified}
}
\description{
Builds a rewardb dashboard for all exposures for subset outcomes or all outcomes for subset exposures
Requires the rewardb results to be generated already.
This exports the data to a db backend and allows the config file to be used to run the shiny dashboard
}
