% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calibration.R
\name{computeCalibratedRows}
\alias{computeCalibratedRows}
\title{Compute calibrated rows}
\usage{
computeCalibratedRows(positives, negatives, idCol, calibrationType = 1)
}
\arguments{
\item{positives}{this is the cohort set that should be calibrated}

\item{negatives}{these are the negative control cohort results}

\item{idCol}{- either target_cohort_id or outcome_cohort_id, this function is used in p}

\item{calibrationType}{- value stored in calibrated column of table}
}
\value{
data.frame
}
\description{
Actual calibration is performed here in a dplyr friendly way
}
