% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildCohorts.R
\name{createCohorts}
\alias{createCohorts}
\title{Create exposures cohorts in the CDM - this function can take a long time}
\usage{
createCohorts(connection, config)
}
\arguments{
\item{connection}{DatabaseConnector connection to cdm}

\item{dataSources}{dataSources to run cohort on}
}
\description{
Create exposures cohorts in the CDM - this function can take a long time
}
