% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildCohorts.R
\name{createCustomDrugEras}
\alias{createCustomDrugEras}
\title{Create custom drug eras}
\usage{
createCustomDrugEras(configPath)
}
\arguments{
\item{configPath}{path to cdm config}
}
\description{
create the custom drug eras, these are for drugs with nonstandard eras (e.g. where doeses aren't picked up by
repeat perscriptions). Could be something like a vaccine where exposed time is non trivial.
deprecated It is now best to use atlas cohort definitions, will be removed in future version
}
