% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildCohorts.R
\name{createOutcomeCohorts}
\alias{createOutcomeCohorts}
\title{Create outcome cohorts in the CDM - this function can take a very very long time}
\usage{
createOutcomeCohorts(connection, config)
}
\arguments{
\item{connection}{DatabaseConnector connection to cdm}

\item{dataSources}{dataSources to run cohort on}
}
\description{
Create outcome cohorts in the CDM - this function can take a very very long time
}
