% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Config.R
\name{.setDefaultOptions}
\alias{.setDefaultOptions}
\title{Sets the default options to the app context.
Will be more widely used in future iterations}
\usage{
.setDefaultOptions(config, defaults)
}
\arguments{
\item{config}{list of configuration options}
}
\description{
Sets the default options to the app context.
Will be more widely used in future iterations
}
