% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExecuteDataBuild.R
\name{getCdmVersion}
\alias{getCdmVersion}
\title{Get Cdm Version}
\usage{
getCdmVersion(cdmConfig)
}
\arguments{
\item{connection}{DatabaseConnector connection}

\item{config}{reward cdmConfig loaded with cdmConfig}
}
\value{
String containing version number or Unknown
}
\description{
Get the cdm version used (e.g. 5.3.1)
}
