% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calibration.R
\name{getOutcomeControls}
\alias{getOutcomeControls}
\title{Returns a set of specified negative controls for input target cohorts}
\usage{
getOutcomeControls(
  appContext,
  connection,
  minCohortSize = 10,
  targetIds = NULL
)
}
\arguments{
\item{appContext}{rewardb app context}

\item{connection}{DatabaseConnector connection to postgres rewardb instance}

\item{minCohortSize}{smaller cohorts are not generally used for calibration as rr values tend to be extremes}

\item{outcomeCohortIds}{cohorts to get negative controls for}
}
\value{
data.frame of negative control exposures for specified outcomes
}
\description{
Returns a set of specified negative controls for input target cohorts
}
