% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Config.R
\name{getPasswordSecurely}
\alias{getPasswordSecurely}
\title{Gets password from user. Ignored if REWARD_B_PASSWORD system env variable is set (e.g. in .Rprofile)}
\usage{
getPasswordSecurely(.envVar = "REWARD_B_PASSWORD")
}
\description{
Gets password from user. Ignored if REWARD_B_PASSWORD system env variable is set (e.g. in .Rprofile)
}
