% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{getUncalibratedExposures}
\alias{getUncalibratedExposures}
\title{get exposures that are not calibrated}
\usage{
getUncalibratedExposures(appContext, outcomeCohortIds)
}
\arguments{
\item{appContext}{rewardb app context}
}
\value{
data.frame of uncalibrated results from databases
}
\description{
get exposures that are not calibrated
}
