% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calibration.R
\name{getUncalibratedOutcomes}
\alias{getUncalibratedOutcomes}
\title{Get uncalibrated outcomes}
\usage{
getUncalibratedOutcomes(appContext)
}
\arguments{
\item{appContext}{rewardb app context}
}
\value{
data.frame of uncalibrated results from databases
}
\description{
get outcomes that are not calibrated
}
