% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Atlas.R
\name{insertAtlasCohortRef}
\alias{insertAtlasCohortRef}
\title{Adds atlas cohort to db reference, from web api
Inserts name/id in to custom cohort table
Maps condition concepts of interest, any desecdants or if they're excluded from the cohort}
\usage{
insertAtlasCohortRef(
  connection,
  config,
  atlasId,
  webApiUrl = NULL,
  cohortDefinition = NULL,
  sqlDefinition = NULL
)
}
\arguments{
\item{connection}{DatabaseConnector::connection to postgres}

\item{config}{rewardb global config}

\item{atlasId}{id to atlas cohort to pull down}
}
\description{
Adds atlas cohort to db reference, from web api
Inserts name/id in to custom cohort table
Maps condition concepts of interest, any desecdants or if they're excluded from the cohort
}
