% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Config.R
\name{loadAppContext}
\alias{loadAppContext}
\title{Loads Application Context}
\usage{
loadAppContext(configPath, globalConfigPath, .env = .GlobalEnv)
}
\arguments{
\item{configPath}{is a yaml file for the application configuration}

\item{globalConfigPath}{path to global yaml}

\item{.env}{environment to load variable in to}
}
\description{
By default, loads the database connections in to this object
loads database password from prompt if REWARD_B_PASSWORD system env variable is not set (e.g. in .Rprofile)
The idea is to allow shared configuration settings between the web app and any data processing tools
}
\examples{
loadAppContext('config/config.dev.yml', 'config/global-cfg.yml')
}
\keyword{appContext}
