% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YamlConfig.R
\name{loadGlobalConfig}
\alias{loadGlobalConfig}
\title{Loads global config}
\usage{
loadGlobalConfig(globalConfigPath)
}
\arguments{
\item{globalConfigPath}{path to global yaml}
}
\description{
Load reward global config yaml file
}
