% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildDashboardDatabase.R
\name{metaAnalysis}
\alias{metaAnalysis}
\title{Perform meta-analysis on data sources}
\usage{
metaAnalysis(table)
}
\arguments{
\item{table}{data.frame}
}
\value{
data.frame - results of meta analysis
}
\description{
Perform meta-analysis on data sources
}
