% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreComputeNulls.R
\name{runPreComputeNullDistributionsJob}
\alias{runPreComputeNullDistributionsJob}
\title{Run pre compute null distributions Rstudio Job}
\usage{
runPreComputeNullDistributionsJob(
  globalConfigPath,
  analysisId = 1,
  sourceIds = NULL,
  nThreads = 10,
  getCemMappings = TRUE,
  minCohortSize = 5,
  workingDir = getwd()
)
}
\arguments{
\item{workingDir}{Charachter working directory (default is currentWd)}

\item{computeExposures}{Compute exposure distributions}

\item{computeOutcomes}{Compute outcome null distributions}
}
\description{
Run task to pre-compute null distributions for all outcomes and exposures as RStudio Job
This process requires a CemConnector connection, this may be slow if using cem.ohdsi.org
}
