% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSccResults.R
\name{runScc}
\alias{runScc}
\title{Peform SCC from self controlled cohort package with rewardbs settings}
\usage{
runScc(
  connection,
  config,
  dataSource,
  exposureIds = NULL,
  outcomeIds = NULL,
  cores = parallel::detectCores() - 1
)
}
\description{
Peform SCC from self controlled cohort package with rewardbs settings
}
